pyproject.toml
[project]
name = "image-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.85",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "django-stubs[compatible-mypy]>=6.0.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.9",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "mysqlclient>=2.2.8",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.9",
]


[tool.ruff]
target-version = "py311"
line-length = 100
[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]

ignore = [
    "E501",   # line too long (handled by formatter)
]



[tool.ruff.lint.isort]
known-first-party = ["image-processor"]

[tool.mypy]
python_version = "3.12"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false # Crucial for gradual typing
disallow_any_unimported = true
no_implicit_optional = true

plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]  # <-- New section for plugin-specific config
django_settings_module = "config.settings"
strict_settings = false


[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "openrouteservice.*"
ignore_missing_imports = true